requirements.txt
Django==5.1.1
environs==14.1.0
djangorestframework==3.15.2
